ECHO_TO_CONSOLE does not count as tracing startup.
Currently we assume we're in startup tracing if the TraceLog has been enabled.
But, in the case of ECHO_TO_CONSOLE we will have started the trace log but we
don't want the startup tracing code to automatically shut down our console
tracing.

This CL makes the is_tracing_start_ flag depend on both IsTracing and that
ECHO_TO_CONSOLE is not enabled.

BUG=trace-viewer:516

Review URL: https://codereview.chromium.org/134183003