[NaCl SDK Docs] Move doc images from "_images" to "images" for Chromesite.
Sphinx will copy the images into the build directory, but it renames the
directory to "_images", which breaks all the links.

This CL modifies the doc Makefile to move the _images directory to the
images directory after running sphinx.

(No trybots as this is a documentation only change)

BUG=none
R=sbc@chromium.org

Review URL: https://codereview.chromium.org/136273005