Adds the new image for display notifications (2nd)
Previous one was reverted due to failure in Windows; I put the
icon images to cros/ directory but some test code may refer to
that file.

That was actually unexpected, those files are ChromeOS only,
so I excluded the code from non-ChromeOS platforms as r246258.

Now this is safe to reland the same patch.

BUG=254737
TBR=oshima@chromium.org, rouslan@chromium.org

Review URL: https://codereview.chromium.org/145013005