Add Windows Flash hw video decode trial.
This is to support the Windows rollout of the feature.

The distribution on different channels will be controlled from the
server. Initially we will target 100% on canary/dev, 50% on beta and
starting slowly (say from 1%) on stable but this will be gradual and
is subject to approval.

Also remove low latency audio trail which has expired a long time ago.
(It also is hard wired to enabled in all Flash binaries.)

BUG=chromium:335836
TEST=Verified Flash flags are properly passed on Link.

Review URL: https://codereview.chromium.org/141753006