Limit apparent duration of decoded file for WebAudio on Android
If the duration of an encoded audio file is unreasonably large, set
the duration to 0, which forces the caller not to preallocate space.
Instead the file will be decoded in chunks and buffered until the
decoding is done. This works around the incredibly huge duration
returned by MediaFormat in the bug report.

This has a side effect that we now end up using twice as much space in
the caller because it has to buffer the data until complete and then
copy it all to the final audio bus.

BUG=326856

Review URL: https://codereview.chromium.org/133583007