Fix the tab dragging issue seen with maximized chrome windows in the recent Canaries.
Turned out to be a fallout from hiding the scrollbars in the OnEnterSizeMove function which is invoked via the message
WM_ENTERSIZEMOVE. This message as the name suggests also comes in for regular moves or window drags. Hiding the scrollbar
lead to badness with subsequent paints.

Fix is to detect if we are in a sizing operation and set a flag indicating the same. The OnEnterSizeMove and OnExitSizeMove
handlers continue hiding the scrollbars and readding the scroll styles as before. The flag is set and reset in OnSysCommand
for sizing operations.

BUG=336325
R=cpu@chromium.org, sky@chromium.org, sky

Review URL: https://codereview.chromium.org/141003016