Ensure LogMessage is called on the correct thread.
ChromeWebRtcLogMessageDelegate was using CalledOnValidThread() to determine
whether to trampoline the request to the IO thread, but CalledOnValidThread()
compiles to "return true" in Release builds (see NonThreadSafe implementation),
so many log messages and IPC requests were happening on the wrong thread.
This CL fixes the bug so that logging always happens on the IO thread.

BUG=335027,335534
TBR=tommi,grunell
NOTRY=True

Review URL: https://codereview.chromium.org/130213010