Check for any open browser windows before clearing session data.
Previously, we used the SessionService to detect when the last browser
window for a profile is closed. With this change, we check for any open
browser windows for the current profile before clearing session data.
This avoids clearing session data when the last normal browser window is
closed but a popup is still open.

BUG=334134

Review URL: https://codereview.chromium.org/139043004