mac: Rewrite some confusing code in web_contents_view_mac.
No intended behavior change.  Functionally, RenderWidgetHostViewMac is the
owner of the RenderWidgetHostViewMacDelegate. The old code failed to reflect
this, and instead leaked a reference to the RenderWidgetHostViewMacDelegate,
and then 'autoreleased' it magically at the right time. This relies on the
assumption that 'RenderWidgetHostViewMac::SetDelegate()' is called at most
once. If it is called more than once, memory is leaked.

The new code is functionally equivalent (with the exception of an 'autorelease'
being replaced by a 'release') but more accurately reflects the ownership of
the RenderWidgetHostViewMacDelegate property. The new code is robust against
multiple calls of 'RenderWidgetHostViewMac::SetDelegate()'.

BUG=NONE

Review URL: https://codereview.chromium.org/133393004