Compute accurate screen space bounds for animated layers.
This walks up the layer tree transforming a layer's bounds as a box
through the different layer transforms and uses the
LayerAnimationController::AnimationBoundsForBox to transform the box at
layers which are animated.

Expansion of the box is included for filters which change the bounds,
however, the LayerAnimationController implementation is currently stubbed.

+ Added computation of animation bounds of layers in the screen space.
- Removed invalid assertion in LayerAnimationControllerTest.AnimatedBounds.
~ Renamed methods to use the uniform naming AnimationBounds.
~ Extracted LTHC::ApplySublayerTransformAboutAnchor from
  LTHC::CalculateDrawPropertiesInternal.
~ Fixed LayerImpl::AsValueInto to correctly calculate bounds.
! Did not substitute implementation of ApplySublayerTransformAboutAnchor
  inside CalculateDrawPropertiesInternal due to performance regression.


BUG=280375

Review URL: https://codereview.chromium.org/109193002