Removes the rotation and scaling in SnapshotAura.
It seems that the compositor already takes care of the screen
rotation and scaling. SnapshotAura doesn't need to do that.

Also, snapshot_aura_unittest.cc is not good for rotation and
scaling because the existing test just paint plain red. This
CL changes the paint content to fail if the rotation is not
handled correctly.

BUG=337559
R=piman@chromium.org, enne@chromium.org, sky@chromium.org
TEST=manually && snapshot_unittests

Review URL: https://codereview.chromium.org/137143005