Rate limit tree selection change side effects
Since tree selection changes triggers async loading of bookmarks
(which cannot be aborted) we ended up with the load coming back in
random order. When the result comes back we update the view and sync
the selection, but the selection might have changed since then.

We now rate limit the side effects of changing the selected item in
the tree and only start loading the bookmarks when the selection
stopped moved.

BUG=336180

Review URL: https://codereview.chromium.org/145983003