Force software for menus to work around corruption in classic
This occurs only in the particular circumstances of running in 
Classic mode, using the GPU rendering path in Chrome, and
opening the menu on a secondary monitor. It appears to be that
when the mouse is moved the contents of the GDI DC are redrawn
by Windows, rather than the surface that was presented by the
D3D path.

To workaround this problem, force the menu to software
compositing. This should be faster than rendering in gpu and
doing a readback to draw on to the GDI surface).

(This is to be merged potentially all the way back to 32.)

R=jbauman@chromium.org, ben@chromium.org
BUG=333380

Review URL: https://codereview.chromium.org/133173005