cc: Make clear the origin of resources in ResourceProvider.
Currently, the delegated resource (allocated by the child compositor) is
regarded as internal, which means the external variable is set to false.
However, it's very confusing because we think internal resources should be
updated by the current compositor.
So this CL introduce Origin enum to identify Internal, External, and Delegated.

BUG=337519

Review URL: https://codereview.chromium.org/145773002