Prevent partial touch sequences from reaching the renderer
Consolidate the various state-related members of TouchEventQueue, instead
relying on but one state variable to determine touch dispatch. In addition,
support touch forwarding suppression per-touch sequence. This prevents cases
where a partial touch sequence might be forwarded, e.g., if a touch handler is
added in the middle of a gesture, or a scroll-ending event is received in the
middle of a gesture.

BUG=295075

Review URL: https://codereview.chromium.org/130993003