cc: Wrap ThreadProxy variables with thread-checking accessors
There's been a number of ASAN-related issues in ThreadProxy with data
races, where we would access members at the wrong time.  This has tried
to be solved with careful variable names and a small number of accessors
(like for LTH).  I think it's time to just wrap everything so that we
can be more confident about thread correctness.

Other than renaming, the code changes here are to pass the
layer_tree_host_id as part of the PostTask rather than accessing a const
variable on both threads.  This is probably moot, but makes it clearer.
Additionally, contents_texture_manager_on_impl_thread_unsafe_ (now
impl().contents_texture_manager) is cleared on the compositor thread
instead of sketchily on the main thread during Stop.

BUG=334611

Review URL: https://codereview.chromium.org/131333003