Add support for closing Android Shell instances from Java.
This reverts commit 4e5242f7faa3f7ced3e8a955a8154c43016986d8.

This differs from the first patch in that creating a new shell does not
clean up the previous instances (i.e. by launching a new window via js).
That had timing dependencies that broke the webkit tests and I believe
the correct solution is to allow switching back and forth to various shells
and allowing the user to delete them.  That will be a larger patch that
I will handle separately.

TEST=webkit/tools/layout_tests/run_webkit_tests.py --platform=android --debug
TEST=build/android/test_runner.py instrumentation --test-apk ContentShellTest --test_data chrome:chrome/test/data/android/device_files -f ContentShellShellManagementTest#
BUG=

Review URL: https://codereview.chromium.org/144473002