Add the scheme chrome-distiller:// and hook up data source.
This adds the scheme chrome-distiller:// and has a minimal implementation
of a data source which should be loaded when a URL with the new scheme is
loaded.

The chrome-distiller:// scheme will be used for displaying distilled
articles, which is extracted content of web pages with long form articles.

Chrome will maintain a list of such articles, and in addition, the user
will be able to display such distilled content on demand for a given URL.

BUG=319881

Review URL: https://codereview.chromium.org/105723002