Set the remote window HWND handle to RemoteWindowTreeHostWin as early as possible so that other parts can depend on it during early initialization.
Since r242975 which changed the initialization order, IME has been
regressed on Metro mode and IME wasn't available in Metro/Ash
mode (Desktop is Okay).

This was because IsRemoteInputMethodRequired() function
relies on the root window HWND which was unavailable
due to the initialization order change, and wrongly
IMM32 InputMethod was instantiated.

The cause was originally investigated and fixed by yukawa@.

BUG=334380
TEST=manual (install, run in Metro Ash mode, activate IME)

Review URL: https://codereview.chromium.org/138843004