Use AboutToNavigateRenderView in prerender PendingSwaps.
Instead of attaching to all created RenderViewHosts, be more targetted and
attach only to the RenderViewHost to be navigated in. This is not directly
related to PrerenderPageNewTab flakiness or beforeunload, but having a more
targetted throttle will making it easier to delay swaps until throttle is
attached, which should resolve both problems.

Tested manually by checking PrerenderPageNewTab* with print statements in
PrerenderPendingSwapThrottle. The relevant tests themselves are still flaky.

BUG=335835,304932

Review URL: https://codereview.chromium.org/144543008