Fix serialization for chrome_settings_overrides manifest permission
The existing manifest permission ID includes a period character which
results in the persisted pref being created with a nested
dictionary. The pref loading code does not expect the nested dictionary
and fails to read the pref.

This change renames the ID to be based on the chrome_settings_overrides
key since the manifest permission as a whole is associated with this
top-level key.

BUG=338505
R=kalman@chromium.org

Review URL: https://codereview.chromium.org/136683008