IndexedDB: fsync after transactions.
IndexedDB corruption rose, only on CrOS, in the 32 release. Included in
that release was a change that stopped calling fsync after transactions 
commit. It is the only IDB change that possibly sacrficed stability to get
better performance. So: revert it and see if corruption jumps back to 31
levels.

BUG=299831,338385

Review URL: https://codereview.chromium.org/137423006