Fix HTML select boxes so that they close when a window is dragged.
This regressed as a result of https://codereview.chromium.org/102253008 because ToplevelWindowEventHandler is now an event handler on the RootWindow instead of an event handler on the window containers and consumes events before they get to EventFilterForPopupExit.

This CL makes EventFilterForPopupExit a pre target handler on the aura::Env object. This closes the select box when the user clicks on the primary or secondary monitor's desktop background as a bonus

BUG=336384
TEST=Manual, see bug

Review URL: https://codereview.chromium.org/143523009