Fix dangling pointer in Isolated FS MessageFilter
In CrxFs API, PepperIsolatedFileSystemMessageFilter used to
const-ref to BrowserPpapiHost via GetProfileDataDirectory.  But when
plugin is dead or dying, it's possible that BrowserPpapiHost has been
destroyed but MessageFilter is still alive, since the former is not
ref-counted.  When that happens, PepperIsolatedFileSystemMessageFilter
holds a dangling reference.

MessageFilter now stores profile directory itself.

TEST=out/Debug/browser_tests --gtest_filter="PPAPITest.*"
BUG=N/A

Review URL: https://codereview.chromium.org/147303002