Upstream the updated Java Omaha XML parser
We hit an issue where the server protocol was updated on 10/29 to send an additional tag was not part of the original spec.  This caused all Chrome clients across all versions to simultaneously think that the server was failing to send a valid response and send an increased number of pings to the server.
    
This CL rewrites the Java parser to be more robust to random changes to the XML as a precursor to the C++ reimplementation (which is still a long ways out).  The code also gets moved upstream so that we can get it reviewed easier.

BUG=181323
TEST=ResponseParserTest

Review URL: https://codereview.chromium.org/67313003