Apps: Grant sufficient permissions to the filesystem opened by fileSystem API
A fileSystem opened by chrome.fileSystem API is passable to a pepper plugin[1],
but currently we cannot open any files with CREATE/TRUNCATE flag under the
filesystem in plugin-side because the filesystem doesn't have sufficient file
permissions (CREATE_OVERWRITE_FILE_PERMISSION is missing).

To fix that, this change calls GrantCreateReadWriteFileSystem() for the
filesystem instead of GrantCreateFileForFileSystem().

(NOTE) This change works fine only when choosing a directory by the API, and
the same issue still happens when choosing a file. I'll address the case in
a separate CL.

[1] http://crbug.com/177017


BUG=336160
TEST=manual (open a file with CREATE and/or TRUNCATE flag under the passed filesystem in plugin)

Review URL: https://codereview.chromium.org/146283002