Performance fixes for CoreAnimation.
Consolidate the calls to [NSScreen screens] because this call can take
several milliseconds, and use an observer to ensure that this call is
only made when the value is changed.

Only use a single OpenGL context because using multiple contexts does
not improve performance and only adds complexity.

BUG=245900

Review URL: https://codereview.chromium.org/134753008