[Mac] Remove a panel from stack window bounds update list when it is being removed
The crash in the bug seems to occur when 2 stacked panels are being removed
in a very short peirod of time. When 1st panel is being removed, other stacked
panels could be queued for update. If 2nd panel, that is in this update list,
is then removed, it should be removed from the update list.

This only occurs on Mac. Windows does not have the problem due to the similar
logic. Linux GTK does not either due to that the update will be performed
immediately.

BUG=336582

Review URL: https://codereview.chromium.org/143453008