Less is more: remove inject_test_data and replace with a simpler, more flexible, method
Basically, add a nullary function 'test_data' to steps. The runloop will call it
to get the default StepTestData for the step. Now we get the benefit of
inject_test_data, but with less magic, and without the requirement that the
annotated method return a single step (as opposed to being a generator, or
returning a list).

This should make life easier for WYK as well.

R=agable@chromium.org
BUG=

Review URL: https://codereview.chromium.org/148153005