Sort issues by their age in the commit queue.
Previously, CQ was processing issues sorted by issue ID, which may result in processing a very recent CL before a very old CL. The age is based on when the Commit box was last checked (and not when the CL was first created).

This will hopefully help to clear the oldest CLs first, improve patch success rate, and keep the CQ's maximum age down.

BUG=335187

Review URL: https://codereview.chromium.org/145293006