telemetry: Compute discrepancy per timeline range and return max.
This patch changes all metrics in RenderingStats to be lists of lists
(one per timeline range). Right now, we only need this feature for frame_timestamps,
but it is cleaner to handle all metrics the same way and it preserves more
information.
The discrepancy calculation was modified to handle list-of-list inputs
correctly.

R=nduca@chromium.org,dominikg@chromium.org
BUG=338796

Review URL: https://codereview.chromium.org/149093003