Copy CygWin1.dll back to its original location in lighttpd distro
In eventual attempt to remove third_party/cygwin. This is a copy
of the DLL for now, because build slaves have a step that copies
the DLL from no_dll\ to the root so until that's removed we don't
want that step to fail.

BUG=123026
R=dpranke@chromium.org, iannucci@chromium.org

Review URL: https://codereview.chromium.org/150653002