Expand PrefHashBrowserTest to ensure unloaded profile initialization produces proper hashes.
This is done by ensuring that only Unchanged stats are generated when explicitly loading the second profile.

This is a regression test for a regression that was about to be introduced in patch set 43 of https://codereview.chromium.org/114223002/#ps3400001 (basically the merge made it such that PrefHashFilter::Initialize -- which was just introduced after writting that CL -- would now use pref_hash_store_ directly to StoreHash where as this is wrong for split prefs; I will make this impossible in a following patch set of the afore-mentioned CL).

The technique used to inspect histograms is akin to the one that was removed from the initial pref_metrics_service_unittest.cc in https://codereview.chromium.org/90563003.

BUG=329078,336478
TEST=New test passes on trunk, but fails on top of patch set 43 of afore-mentioned CL.

Review URL: https://codereview.chromium.org/137383011