WebMediaPlayer{Android,Impl}: use RenderFrameObserver instead of ML::DestructionObserver/RenderViewObserver.
RenderThreadImpl is torn down before the main thread's MessageLoop is, and some
of WebMediaPlayerAndroid's dependencies require the render thread to still be
alive, so move its tear-down to earlier in the process.
This is the Android version of r240125 which fixed the same bug in
WebMediaPlayerImpl, but forgot that WebMediaPlayerAndroid is a fork.

Since RenderViewObserver is apparently out of favor nowadays, use RenderFrameObserver.

BUG=338393,304967,338481

Review URL: https://codereview.chromium.org/137243005