[Files.app] Convert a stored shortcut path to a real path with multiprofile-aware mounted path.
The mount path of the drive is currently fixed (/drive) but will be changed to the multiprofile-aware mounted path. But we decided the keep the stored shortcut paths with the previous mounted path (/drive) for compatibility.

This patch adds the logic to changed the fixed '/drive' to the current mounted path. Until multi-profile is introduced, the drive path is still '/dvive', so this does nothing (converting '/drive' to '/drive' to be exact).

See the bug for detail of multi-profile.

BUG=336123
TEST=manually tested.

Review URL: https://codereview.chromium.org/145493003