Dispatch chrome.app.runtime.onRestarted to ephemeral apps
If Chrome is restarted while an ephemeral app is running, the app will be dispatched the onRestarted() event on restart. If the app does not listen to this event, it is just relaunched.

BUG=339001
TEST=To test dispatch of the onRestarted event:
Find the "Restarted Event Sample" app in the Web Store and run it
ephemerally. Restart Chrome, e.g. using chrome://restart, and
verify that the app is restarted.

Review URL: https://codereview.chromium.org/133073014