android envsetup: Stop explicitly setting host_arch when target_arch=x86.
The explicit setting seems to match the default value for host_arch in
build/common.gypi close to line 110, and it's not set explicitly for other
target_archs, so this looks like an oversight. It's been this way since this
file was added in r148790 18 months ago.

No intended behavior change.

BUG=330631

Review URL: https://codereview.chromium.org/142023006