Reduce spam from CQ.
My previous change https://codereview.chromium.org/138173005/ made CQ
post a message when the commit box was unchecked.  It shouldn't have
posted it when CQ committed a change, but for some obscure reason it
did.  Sometimes twice...

This patch unifies the way an issue is removed from the queue, and in
particular, should eliminate most of the unwanted messages.  Some CQ
messages about unchecked CQ box will still be posted when CQ drops it,
and it is expected to be rare.

BUG=339116
R=phajdan.jr@chromium.org

Review URL: https://codereview.chromium.org/130983010