Add the Bot Update step
This step is added to all gclient update as an escape hatch for us to partially switch from updating from SVN to updating from Git, make script changes without master restart, and activate/deactive git mode without master restarts.

What will happen in a future CL is that the bot_update.py script deposits a flag file that tells gclient.py and apply_patch.py to not run if it detects that flag.

For this CL, the requirements are that:
1. This adds a "bot update" step to all builds
2. The bot update step should be a no-op
3. the bot update step should have all information needed to get the correct checkout.

BUG=333575
TEST=Ran locally

Review URL: https://codereview.chromium.org/147413003