Update WebNode::getElementsByTagName() callers to use a WebNodeCollection
Update WebNode::getElementsByTagName() callers to use a WebNodeCollection
instead of a WebNodeList. The WebNode::getElementsByTagName() public API was
updated in Blink r166263 to return a WebNodeCollection instead of a
WebNodeList.

Callers need to be updated so that we can get rid of the current workaround
allowing a WebNodeList to be implicitly converted into a WebNodeCollection.

BUG=235008

Review URL: https://codereview.chromium.org/152133004