Move files from //chrome/browser/sync to sync_driver component.
This CL moves files that depend on the now-componentized data_type_controller.h
and data_type_manager.h into the sync_driver component. For the most part these
files are moved as-is. The only exception is sync_model_association_manager,
which had a DCHECK about being on BrowserThread::UI; this DCHECK is removed.

BUG=339726
TBR=joi

Review URL: https://codereview.chromium.org/133503011