Transfer appcache state during cross site navigations too.
The AppCacheHost is an object that persist across the lifetime of a document. It's created prior to the document and prior to the initiation of the resource load for the documents main resource. Today, when a cross-site xfer occurs the host is deleted and a new host is created that knows nothing of whats going on... so not surprisingly things are broken.

Either state of the old host that gets deleted needs to be saved and restored into a new host that gets created... or the old host should not get deleted and instead get transferred along as well.

I chose the latter because there's some state that can't be saved/restored so easily and moving the whole object is less tedious.

BUG=331270

Review URL: https://codereview.chromium.org/137883005