Makes Wiget::OnNativeWidgetMove() be invoked during a maximize
A maximize animation triggers recreating the layer of the
window. Recreating the layer forces the bounds of the window to 0. A
maximize results in the origin going to 0. Together this meant a
maximize from any where resulted in the code not thinking a move
happened.

This patch special cases NativeWidgetAura::OnBoundsChanged to assume
that if the old bounds was empty at 0,0 and the new bounds is not
empty that a move happened.

BUG=335486
TEST=covered by test
R=ben@chromium.org

Review URL: https://codereview.chromium.org/153803006