[Android WebView] Fix small licenses-related inconsistencies
1) Add Apple PSL 2 (APSL 2) license into the whitelist.

2) Fix breakpad's "License" field to be "New BSD, Apple PSL 2.0 and Apache 2.0".
Dropped "MIT/X derivate" because I've found no mentions of it in the project.

3) Fix webdriver's "License" field to be "Apache 2, MIT and GPL v2". The latter
corresponds to old jQuery code, which was previously licensed under GPL v2, not
GPL, as the field was stating.

These changes make "breakpad" and "webdriver" projects to be recognized as
"compatible" with the Android tree.

R=frankf@chromium.org, mkosiba@chromium.org, thestig@chromium.org

Review URL: https://codereview.chromium.org/135533007