Use gyp targets to link in X11 libraries instead of setting -lXfoo
This sets up gyp targets for the various X11 libraries we use and uses
pkg-config to determine the correct cflags/ldflags incantations to use to
link against them instead of listing out -lXfoo lines in libraries sections
in gyp targets throughout the tree.  This is more likely to generate the right
lines and makes the dependencies a bit more explicit.

BUG=339518

Review URL: https://codereview.chromium.org/153223002