Fix TabCaptureApiTest.GrantForChromePages flakey failures
In this test we base::Bind DidCopyFromBackingStore into callbacks to be
executed OnSnapshot. If we WebContentsCaptureMachine::Stop before these
callbacks come in, then render_thread_ will be set to NULL, and if
these callbacks end up being executed after this point, but before the
WebContentsCaptureMachine is destroyed, access to to a NULL
render_thread_ will result.

BUG=340475

Review URL: https://codereview.chromium.org/151733003