WebRTC browser tests will now render during tests.
This is a speculative fix to try to fix the video quality test timeouts
on the FYI bots. I noticed the browser tests did not run with pixel
output enabled, which means they don't really render. Not sure how it
is related to the gUM failures in the bug, but this will at least make
it easier to see what's going on when it fails.

R=xians@chromium.org
BUG=340609
NOTRY=True

Review URL: https://codereview.chromium.org/153623005