DeclarativeContentApiTest.Overview shouldn't rely on details of layout timing
This CL fixes the DeclarativeContentApiTest.Overview browser_test not to depend
on exact details of when Blink schedules layout. Instead of taking two trips
through the event loop, we just force a synchronous layout.

This CL fixes a bug that will be uncovered once we move layout to be driven off
requestAnimationFrame instead of off a timer.

R=jyasskin@chromium.org
BUG=340357
NOTRY=true

Review URL: https://codereview.chromium.org/154443002