Attempt at fixing possible shutdown crash
In some situations we can enter ~BrowserView with WebContents in the
tab strip. There is code that attempted to deal with this, but by the
time we get to ~BrowserView it's really too late to attempt to delete
the WebContents. Reason being doing anything with the TabStrip more
often than not results in calling back to the BrowserView and related
classes. So, this shutdown really needs to be done earlier on.

I'm moving it to when the frame is being destroyed, which means all
view/layout classes are still valid.

BUG=176556
TEST=covered by test now.
R=ben@chromium.org

Review URL: https://codereview.chromium.org/154633003