Pepper: Remove ErrorInfo from SelLdrStartParams.
This cleans up the usage of ErrorInfo to shorten its scope in several places.
This will make life easier during the trusted plugin refactor, when
establishing the connection between the plugin and renderer has to be made
asynchronous.

This also makes SelLdrStartParams not have any members that are written to,
simplifying how the struct is used.

BUG=333950

Review URL: https://codereview.chromium.org/155863004