Rename and unify paths for LayerTreeTest::CreateFakeOutputSurface.
This was called CreateFakeOutputSurfaceForTest because of the conflict
with another CreateFakeOutputSurface method. To avoid this confusion,
replace the CreateFakeOutputSurface() calls with direct calls to
FakeOutputSurface::Create3d().

Get all unit tests that use a FakeOutputSurface (everything except
pixel tests) to override CreateFakeOutputSurface instead of
CreateOutputSurface. This gives us a common point for all unit tests
to inject things onto the OutputSurface. We'll inject the frame rate
there in the future (moving it out of LayerTreeSettings).

R=jamesr
BUG=267140

Review URL: https://codereview.chromium.org/155343002