aura: Don't leak GLHelper when destroying causes new one to be made.
When we destroy the GLHelper we cancel all its current tasks. If one
of those tasks is a yuv readback for RenderWidgetHostViewAura, it
ends up in RWHVA::CopyFromCompositingSurfaceFinishedForVideo. This
method tries to get a GLHelper in order to insert a sync point for
giving back to the compositor.

Since we're inside the destructor of the GLHelper held in the
GpuProcessTransportFactory, getting a new GLHelper fails to store
the new GLHelper* in the gl_helper_ member, and the GLHelper* is
lost, leading to a leak.

R=piman
BUG=337153

Review URL: https://codereview.chromium.org/149993004