Provide optional TouchMove suppression within a slop region
Add support for suppressing TouchMove events within a slop region, as long as
the touch sequence has not been preventDefault'ed and there is but one touch
pointer in the sequence.  This is a direct port of the current slop suppression
functionality from ContentViewGestureHandler to the TouchEventQueue, and will
be enabled only on Android.

BUG=295075

Review URL: https://codereview.chromium.org/130983005