Render process count should not be forced from java unconditionally
content::SetContentCommandLineFlags should enforce a maxiumum value from java.
The problem is that it overrides the value even if the default would be less.
Instead we should let RenderProcessHost decide it based on the available physical
memory and only override if the default value is too high.

Review URL: https://codereview.chromium.org/138313004