Pass chrome schema to policy::BuildHandlerList()
Chrome schema is necessary to be accessible in policy::BuildHandlerList() since some of the handlers will be migrated to a new policy handler based on schema(as defined in policy_templates.json).

This CL changed the signature of BuildHandlerList() and thus modification to constructor of BrowserPolicyConncector was made, taking callback to BuildHandlerList instead of the result returned by it, and postponed the generation of policy handler list. 

BUG=258339

Review URL: https://codereview.chromium.org/147033004