sync: Apply updates passively without routing info
This CL allows the sync scheduler to make decisions about whether
updates should be applied directly from the sync thread (GROUP_PASSIVE)
or if the work should be delegated to the model thread.  This means
that the sync backend no longer needs access to the routing info.

In a way, this feature is nothing new.  It would have been wrong for the
scheduler to have non-GROUP_PASSIVE routing info when it's in configure
mode, and equally wrong for it to have types in GROUP_PASSIVE when in
normal mode (with the exception of a few control types).  This CL trims
redundant information from the system.

BUG=278484,336616

Review URL: https://codereview.chromium.org/145993004