Remove MediaStreamDeviceThread in favor of audio worker TaskRunner.
Part Deux!  See https://codereview.chromium.org/17444005/ for the
previous attempt.  It was reverted oh so long ago because on OSX we
use the UI thread as the "audio thread" for http://crbug.com/158170.

Now that the AudioManager has the concept of a worker thread for
heavy tasks, this can safely be relanded without fear of blocking
any browser threads.

See http://crbug.com/256945 for more context on the previous revert.

BUG=none
TEST=unittests pass.

Review URL: https://codereview.chromium.org/154143002