Make the member |is_cursor_visible_| in DesktopWindowTreeHostWin static.
On Windows we can have multiple root windows and the implementation
of ::ShowCursor() is based on a counter, so making this member static
ensures that ::ShowCursor() is always called exactly once whenever the
cursor visibility state changes. This prevents an incorrectly hidden
(or incorrectly shown) cursor because the number of calls to
::ShowCursor() is no longer dependent on the number of instantiated
root windows at the time when a request to change the cursor visibility
is made.

BUG=317496

Review URL: https://codereview.chromium.org/151563004