Omnibox: Turn on Reorder Mode By Default
Reorder mode was evaluated as a 30% experiment in Beta.  Based on the
results ( see crbug 273149 ), we've decided to launch it.  We're going
to do a staged rollout.  This change enables reorder mode by default.
We're going to use the variations server to limit reorder mode to (for
now) be on for 90% of users.

BUG=339808

Review URL: https://codereview.chromium.org/147123009