[WebMIDI] Increment in-flight bytes in front of sending data.
If we call AccumulateMIDIBytesSent in MidiManager::DispatchSendData,
it tries to acquire |in_flight_lock_| which is already acquired.
Since recursive locking is not supported by base::Lock in general,
this may result in a deadlock.

This CL fixes it by incrementing |sent_bytes_in_flight_| and releasing the
lock in front of calling DispatchSendData.

BUG=303596
R=toyoshim@chromium.org

Review URL: https://codereview.chromium.org/154153003