sync: Fix auth error test timeouts on trybots
The auth error test are failing because the condition they were waiting
on is ProfileSyncServiceHarness::AwaitCommitActivityCompletion(), but
the commit activity never completes when these tests are simulating an
auth error.  All sync activity stops when the auth error is encountered.

This CL fixes the error by creating a custom StatusChangeChecker for the
sync auth tests.  The tests will now try to create a bookmark, then wait
for *either* the commit to go through or an auth error (specifically, an
error fetching the OAuth refresh token).  Before, only the bookmark
creation case would cause the test to exit.  Once this waiting is done,
the test's helper function will return either true or false based on
whether or not an auth error was detected.

We might want to revisit this when StatusChangeCheckers are augmented to
support failure conditions.  We should fail the test immediately if the
bookmark was committed when we were expecting an auth error, and
vice-versa.  However, the current code is sufficient to fix this bug,
and doesn't touch too many lines of code.

This bug was not a problem on the waterfall.  See the linked bug for an
explanation of why this only affected the trybots.

BUG=341215

Review URL: https://codereview.chromium.org/156383002