Add AppListModel::MoveItemToFolderAt
This is necessary to allow items to be moved to a specific locaiton in
a folder (or in the top level model). 

This also updates MergeItems() to support source items inside folders.

This includes some cleanup to use scoped ptrs to document ownership of
items in AppListModel.


BUG=335761
For app_list.cc
TBR=jamescook@chromium.org

Review URL: https://codereview.chromium.org/152713003