Cleanup time_format.cc, mostly cosmetic, one minor fix
This is a follow-up to the TimeRemainingShort() --> TimeDurationShort()
rename.  It renames some internal variables to match the new publicly
visible function name and consistently orders all occurrences of the
different time formats.  That way, a missing case in one of the switch
statements was found and corrected.

BUG=none

Review URL: https://codereview.chromium.org/155843003