Create a visible window with class name Chrome_RenderWidgetHostHWND which corresponds to the bounds of the web contents.
This is done to ensure that legacy drivers for trackpoints/trackpads and others which use track gestures for back forward
navigations work. These drivers call the WindowFromPoint API and get the process name. For Chrome they look for the old
Window class Chrome_RenderWidgetHostHWND. They use this to determine whether to send mouse wheels and special keyboard
message for back forward navigations, etc.

With Aura we don't have this window class anymore causing Chrome to not work well for a number of users.

To workaround this in the shortterm, a dummy HWND with the Chrome_RenderWidgetHostHWND class name is created. This window
is visible and handles mouse events. It is transparent to ensure that it does not need to draw. To ensure that we don't
mess wit mouse events too much we set capture to the parent window as needed. This window also serves as the parent to
windowless NPAPI plugins like Flash. Additionally the dummy accessibility window with this class name has been removed
and the functionality has been consolidated into the newly added LegacyRenderWidgetHostHWND class which provides this
functionality.

BUG=326022,335941
R=cpu@chromium.org, scottmg@chromium.org, sky@chromium.org, cpu, scottmg
TBR=jam

Review URL: https://codereview.chromium.org/151083002