[OriginChip] Hide/show the origin chip as the Omnibox gains/loses focus.
The ultimate implementation may be rather more complicated, but this simple approach covers nearly all of the currently specified behaviors and should be a good starting point for everyone to play around with.

Also add a check for editing in progress when the browser view sets focus to the location bar so that Ctrl-l/Alt-d won't clobber any edits in progress.

BUG=338563

Review URL: https://codereview.chromium.org/143043004