Add unit test target for shell_dialogs module.
This is necessary to avoid a circular dependency that on Mac is forbidden:

ui_unittests -> shell_dialogs -> aura -> ui_unittests

The down side is that momentarily we lose automatic test coverage on the bots for the shell_dialogs.
This can be fixed by teaching the buildbot scripts to run it after this lands.

BUG=331669, 35878, 299841
TEST=shell_dialogs_unittests
R=ben@chromium.org

Review URL: https://codereview.chromium.org/151253004