Inject default libraries to all targets in GN.
In GYP Mac, most targets got a default set of mac libraries by virtue
of depending on base, which due to a GYP bug pushed the libraries to
all targets depending on base, even separate shared libraries.

In GYP Windows the set of default libraries on Windows is global and
this keeps the same list and behavior.

The lists of libraries could use some cleanup, btu this patch just
attempts to keep the status quo in a cleaner way rather than fixing the
unreasonable list of libraries (especially on Windows).

R=scottmg@chromium.org, scottmg

Review URL: https://codereview.chromium.org/131513026