[cc] Fix for hidden layers with copy requests getting composited to the screen.
In LayerTreeHostCommon, when computing the drawing properties, a layer was being treated as visible if it was hidden and had copy requests.  The intention is for the layer+subtree to be drawn as if the target layer were visible, just for read-back.  However, a distinction must be made between whether a layer is "visible" versus "hidden, but still drawn."  This is important when setting the render surface's contributes_to_drawn_surface property.  If a layer is hidden, it does *not* contribute, regardless of whether it is drawn.

BUG=341318
TEST=New cc_unittests that fail before this change: LayerTreeHostReadbackPixelTest.ReadbackHiddenSubtree_*  Confirmed LayerTreeHostCommonTest.SubtreeHidden*, and all other cc_unittests still pass.

Review URL: https://codereview.chromium.org/142443009