POSIX: replace alarm() with a waiting thread.
Using alarm(2) is very error prone. It's not compatible with certain
implementations of sleep(3) and it's hard to do anything in the handler
because of async-signal safety.

We replace alarm(2) with spawning a thread that will signal the process
after waiting.

Note: this CL only affects a --child-clean-exit codepath.

BUG=341147
NOTRY=true

Review URL: https://codereview.chromium.org/142453003