Make LayerTreeHostClient::Animate take TimeTicks instead of double
This makes LayerTreeHostClient::Animate take a TimeTicks instead of a
double, moving the conversion to double from the caller of Animate to
wherever doubles are actually needed.

This avoids precision loss caused by converting from TimeTicks to
double and back again when TimeTicks are actually needed (as in
LayerTreeHostClientForTesting::Animate) fixing the suspected cause
of flakiness in LayerTreeHostAnimationTestFrozenAnimationTickTime.

R=danakj,jamesr
TBR=sky

BUG=341517

Review URL: https://codereview.chromium.org/153633006