Fix a crash in content-shell headless mode with --dump-render-tree
Currently the platform specific state is not cleanedup in headless mode and
that leads to a crash upon content-shell exit.

This patch deinitialize the platform specific state by calling PlatformExit()
just before discarding the main window so that it will delete the compositor
on exit.

BUG=338919

Review URL: https://codereview.chromium.org/145633009