Allow embedders to control which fullscreen player to release when exiting fullscreen.
The main issue being fixed is that the embedded player does not reflect playback status on remote player. This is because when fullscreen is exited, the fullscreen player's resources are released (which unintentionally releases the remote player's resources since the remote player is the fullscreen player at this point).

Related downstream changes in CL 153495.

BUG=298231

Review URL: https://codereview.chromium.org/130173011