Re-record mobile-news page for key_mobile_sites page set.
We had some issues with Telemetry scrolling the page before it is ready to
scroll. The page now has a variable (NEWS_telemetryReady) to indicate that it's
ready to be scrolled.
This patch contains the recording of the new version of the page and updates the
navigate steps accordingly.

BUG=248928

Review URL: https://codereview.chromium.org/157483002