Do not show 'Minimize' item if the window cann't be maximized.
This is because WidgetDelegate does not have a CanMinimize() method and
some code depends on this to make sure that the minimize item isn't
shown when it should not. The right fix would be to introduce a
CanMinimize() method in WidgetDelegate but we need to fix the regression
first.

BUG=341010

Review URL: https://codereview.chromium.org/135553006