Fixes chromium build when archiving the debug nacl_irt
Now that tls_edit is used as a post process to fix up the TLS for the
nacl_irt, we also need to run it on the debug version of the nacl_irt
so that we can archive away a working version of the debug nacl_irt.

This change is needed to fix the build bots when archiving
the nacl_irt.

R=mcgraphr@chromium.org, sbc@chromium.org
BUG=https://code.google.com/p/nativeclient/issues/detail?id=3127
NOTRY=true

Review URL: https://codereview.chromium.org/153343004