[PPAPI][MediaStream] Rename AudioFrame to AudioBuffer
For audio, a frame usually has the same meaning as sample,
so AudioFrame is not a good name for a bunch of audio
samples. Change it to AudioBuffer.

This CL only changes the name in base classes. The API
interface will be changed in a separate CL.

BUG=330851

Review URL: https://codereview.chromium.org/142023008