Adding application name message in kiosk mode after user timeout
This is a security feature for a user who interacts with a kiosk device - making sure that he/she knows which application is running.

When a user interacts with the device after 2 minutes of inactivity, a message shows which application is running and who was the author.

When spoken feedback is turned on, the device will also narrowate the content of the message.

The message itself will slide in from the top and - after a short timeout - will automatically slide out again.

BUG=268453
TEST=unit test & visual test

Review URL: https://codereview.chromium.org/131353007