Split preserve3d into: "should-flatten" and "is-3d-sorted"
Preserve3d has implications for both sorting and flattening of
transforms. This is a bummer. Sometimes we want to allow a transform to
remain unflattened, but we don't want that to affect 3d sorting. With
this CL, we will have the ability to do that.

Note: this is essentially this cl
https://codereview.chromium.org/100393005/

..minus the generalizations to sorting that were being attempted in that
patch.

Note, this CL is gated on https://codereview.chromium.org/98373011/ 

R=enne@chromium.org
BUG=338980

Review URL: https://codereview.chromium.org/147833003