Custom Downloads path for testing code must be injected via VolumeManager.
This is for bug 336124, where VolumeManager wants to manage all the mounted
path to keep track of the whole list of own profile's directories.
In addition, by this change we can add ForTesting() suffix for the method
that clearly prohibits other types of uses.

BUG=336124
TBR=sky

Review URL: https://codereview.chromium.org/151053005