Attempt to fix crash in base::android::MethodID::Get().
First attempt to land this CL can be found in https://codereview.chromium.org/156353003/. I did a mistake which Ami found after I had committed. The first version was therefore reverted and this is the second attempt to land; this time without the bug that Ami found.

TBR=bulach
BUG=336600

Review URL: https://codereview.chromium.org/152783005