[Mac] In +[CustomFrameView load], do not perform any work if the process has a --type flag.
This flag indicates that it is a non-browser process and so does not need to
use NSWindow. By interacting with the NSThemeFrame class, its +initialize method
runs. This initializer communicates with the WindowServer. Since +load is run
from dyld as a module initializer, this effectively connects to the WindowServer
via static initialization.

BUG=306348
R=mark@chromium.org

Review URL: https://codereview.chromium.org/157763002