Don't send VSync parameters to the host on Mac.
This was broken for a long time, and it looks like fixing it might have
cause some performance regressions. This is to test if that's the case.

If this does not fix the performance issues, I will remove the display
link entirely. If that does not fix the issue, then this is a misattribution.

If this does fix the issue, then we'll need to discuss how vsync impacts
compositor performance, and the validity of the previous baselines.

BUG=341811
TBR=kbr@chromium.org

Review URL: https://codereview.chromium.org/156533005